chore: 프로그램 시작점을 main.js에서 App.js로 바꿈
