+++ index.html
@@ -10,6 +10,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.js"></script>
+    <script type="module" src="/src/App.js"></script>
   </body>
 </html>
